Updated vagrant file
If the files id_rsa.key and id_rsa.pub are present in the tarball in the same directory as the Vagrantfile then they are provisioned over to the newly created virtual machine. Thereafter, the bootstrap.sh will ensure that the copied over files are used rather than newly created key files.